# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Cafeteria Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (own orders for users, all for admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, completed, cancelled]
        - name: tableNumber
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order from cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # PAYMENTS
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate payment for an order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm payment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfirm"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/order/{orderId}:
    get:
      operationId: getPaymentByOrder
      tags: [payments]
      summary: Get payment for an order
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ADMIN DASHBOARD
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderInput:
      type: object
      required: [paymentMethod]
      properties:
        tableNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [cash, card, upi]
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
    OrderItem:
      type: object
      required: [id, productId, productName, price, quantity, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
    Order:
      type: object
      required: [id, userId, status, total, items, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        tableNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [cash, card, upi]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        estimatedReadyTime:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # PAYMENTS
    PaymentInput:
      type: object
      required: [orderId, method]
      properties:
        orderId:
          type: integer
        method:
          type: string
          enum: [cash, card, upi]
        upiId:
          type: ["string", "null"]
    PaymentConfirm:
      type: object
      required: [transactionId]
      properties:
        transactionId:
          type: string
    Payment:
      type: object
      required: [id, orderId, method, status, amount, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        method:
          type: string
          enum: [cash, card, upi]
        status:
          type: string
          enum: [pending, paid, failed, refunded]
        amount:
          type: number
        transactionId:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ADMIN DASHBOARD
...