# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Cafeteria Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  /** @nullable */
  tableNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  paymentMethod: OrderPaymentMethod;
  paymentStatus?: OrderPaymentStatus;
  /** @nullable */
  estimatedReadyTime?: Date | null;
  createdAt: Date;
  updatedAt?: Date;
}